backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
sentence-transformers
groq
pydantic
python-docx
PyPDF2
pycryptodome
requests
